Refactors OpenAISwift slightly to allow not only for custom endpoint providers but also custom request handlers/creation. Moves Config to URLSessionRequestHandler.
